requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
hdbscan==0.8.40
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyaml==25.1.0
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.1
umap-learn==0.5.7
